Add additional typing hints (#53)
* use labeled tuple elements for Point type

Using labeled tuple elements allows editors to show what values are expected in the different positions of the tuple.

* use mapped object type for Unit/Factors

Using a mapped object type allows the acceptable units names to be shared between the Factors type and the arguments to fromTile and the constructor.  This gives good editor support for the acceptable options to these methods.

* use labeled tuple elements for BBox type

Using labeled tuple elements allows editors to show what values are expected in the different positions of the tuple.